+++ packages/guide/src/app/shared/components/item-card-switch/item-card-switch.component.ts
@@ -1,6 +1,6 @@
 import { Component, ContentChild, input, TemplateRef } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CustomEntry, MinimalTagBasedItem } from "@ci/data-types";
+import { CustomEntry, MinimalTagBasedItem, Quality } from "@ci/data-types";
 import { SharedModule } from "../../shared.module";
 import { DatabaseItemDetailsComponent } from "../database-item-details/database-item-details.component";
 import { DatabaseItemDetailsDirective } from "../../directives/database-item-details.directive";
@@ -18,6 +18,9 @@ export class ItemCardSwitchComponent {
     item = input.required<ItemEntry>();
     @ContentChild(TemplateRef) databaseItemDetails: TemplateRef<any> | null = null;
     context = input<ToDoContext | undefined>();
+    amount = input<number | undefined>();
+    quality = input<Quality | undefined>();
+
     isTagBasedItem(item: ItemEntry): item is MinimalTagBasedItem {
         return 'key' in item;
     }